Mejoras en el random

+++ DeathTokenBase/DeathToken/Fighter.h
@@ -41,17 +41,17 @@ public:
     inline string getDesc() const { return desc; }
     inline float getAbility() const { return health + attack * 1.3f; }
     inline int getMaxHealth() const { return maxHealth; }
-    inline int getMindset() const { return mindset; }
+    inline float getMindset() const { return mindset; }
 
     inline bool isAlive() const { return health > 0; } // es un metodo util para saber si el combate sigue o no
 
     // Métodos Setters
     inline void setHealth(int value) { health = value; }
-    inline void setMindset(int value) { mindset = value; }
+    inline void setMindset(float value) { mindset = value; }
 
     inline void takeDamage(int value) { health = health - value < 0 ? 0 : health - value; }
-    inline void boostMindset(int value) { mindset += value; }
-    inline void reduceMindset(int value) { mindset -= value; }
+    inline void boostMindset(float value) { mindset += value; }
+    inline void reduceMindset(float value) { mindset -= value; }
    
 private:
     string name;
@@ -59,7 +59,7 @@ private:
     int health;
     int attack;
     string desc;
-    int mindset;    // Valor adicional que puede afectar el rendimiento
+    float mindset;    // Valor adicional que puede afectar el rendimiento
 };
 
 #endif // FIGHTER_H
